Added a fast path to stamp the sheets using pdftk.
The code now will make use of pdftk if it is available to put the
questionnaire into the background of the marks that are generated by
python. This is much faster than using the python PDF library.

pdftk is still not needed. If it is not available the slower code path
will still be used.