Fix choiceitemtext mostly.
It is not quite working if it is on the same line with other items. In
that case errors occur, and the baseline is not moved up for the other
items.
The baseline needs to be fixed either by moving the other items up too,
or by somehow moving the box down instead of the label up.
The errors, no idea right now.