Implement dynamic paper sizes.
This commit fixes dynamic paper sizes for the ID boxes. This means that
SDAPS will just do the right thing if eg. given a letter sized input document.

What will not work properly is if the paper is not wide enough to fit the boxes,
you can disable them in that case though.