Use draft mode by default; override options during SDAPS setup.
When creating the LaTeX document an internal draft mode is enabled
by default. A file (sdaps.opt) is parsed if it exists, which enables
the sdaps main program to set options and override anything that the
user changed.
This way we can disable all the marks, and enter final mode automatically.