Add more features to the C library.
The following functions are now available externally:
 * get_coverage_without_lines
 * get_white_area_count

The following internal features are added:
 * surface copy utilities
 * hough transformation
 * flood filling
 * thinning algorithm (disabled)
 * kfill algorithm (disabled)
 * debug PNG output routine (disabled)

and some more helper routines.

Another change is the "barrel" loader, which improves loading speed a
little bit (though likely it would be saner to cache the TIFF handle,
instead of doing these kind of micro optimizations).