Rework checkbox detetion and add quality factor.
This commits changes the checkbox detection a lot. The result may be
currently worst then before, but I do think that the change still
makes sense to push out.

So we have now a new "metrics" feature, where the checkbox can be
analysed using different methods, and the best one will be picked
to decide the state. The approach is pretty simplistic, but at least
allows adding more "metrics" to decide the checkbox state based on.

Currently three "metrics" are implemented:
 * coverage, which now only uses the inside of the checkbox
 * cov-lines-removed: First removes the three most important
   lines from the image (using a hough tranformation) and then
   returns the coverage.
 * cov-min-size: Flood fills the image, and returns the white area,
   but only uses areas that exceed a certain size.

To say the truth, the last metric seems to be pretty useless, might
work better if it was filtered before.

I am actually thinking that a new metric that somehow decides how
distributed the black area is would be good. I am thinking that
some statistical analyis of the data or even a transformation into
fourier space could work.