Allow setting oneside in LaTeX. Fix paper size handling.
This changes the default option handling a bit. I am not entirely
sure that this is correct, but it seems to work fine. It is now possible
to change the size of the paper, which would not work before apparently.

Note that the "10pt" option was removed, because it seemed like it had
no effect before.