Some backward compatibility. Survey ID in classic mode is unchanged.
This changes the survey-ID calculation to be the same as before the new
options were added (style, duplex). If both of them are the default value
then they will not be hashed into the Survey-ID.

Another change is a little bit of backward compatibility when loading
projects. Though it does not help much, because the quality value will
be missing in any project that is not very new.