Do not skip reading the barcodes if the matrix fails.
This commit changes the behaviour to not skip reading barcodes when the
matrix recognition fails. This means there are a lot less messages, and
it may help with a better fallback mode.
Obviously this only works in code128 mode, and not in classic mode. This
means that in classic mode the same errors will still be thrown.