Add --force option to add command.
This command can be used when the scan is missing some
pages. It should only be used when one is sure that the scan
is correct other than some pages that are missing. I have seen
scans where a front/back page of a duplex scan was missing,
which would entirely break internal assumptions!