Add get_size function that can use both the 1bpp and rgb surface
This fixes a performance bug in the GUI for images where no matrix was
recognized. Because the fallback matrix is recalculated all the time,
and the A1 surface was not loaded, it was reloading the image for every
scroll event.
As the RGB surface is already loaded in the GUI, we can simple use that
instead.