Add option to export PNG images of freeform fileds.
This commit adds the --images option to the "csv export"
command. If enabled, images will be exported into a subdirectory
with the same name as the file. If the field contains data
then the path to the PNG image is included in the CSV file.
Otherwise a 0 is written out.
Unicode text is still prefered if it exists.