Add new checkmode option
This option allows changing the recognition mode from the
mode where a checkmark is considered checked, and filled is
considered corrected (unchecked).
The two new modes are "check" and "fill" where "check" simply
removes the correction feature, while for "fill" a simple
checkmark might not be enough for detection.

There is no info text yet, so the class errors out if [noinfo] is
not set.

Fixes issue #35.