Do not use g_malloc0_n to be compatible with old glib versions.
Apparently under some setups on OSX ancient versions of GLib might
be loaded (2.22 in this particular case). This patch fixes loading
on this versions by replacing the g_malloc0_n call with g_malloc0.

This patch is save right now, but is incorrect as integer overflows
might occur.
as overflow checking is not really required in this case (an even
larger image would be needed).