Rotate image for conversion based on paper size (fixes #88)
When importing images and converting them to the internal image format
SDAPS assumed that the image should be in portrait mode after conversion.
This only makes sense if the survey itself is portrait, so detect the
portrait/landscape mode of the survey and rotate the image accordingly.

With this commit the landscape/portrait mode of the image has to match
the papers mode. If not, the image will be rotated by 90 degree, which
makes detection impossible.