utils: Fix LaTeX TEXINPUTS handling regression
The commit a380a88f4 (utils: Try to keep TEXINPUTS environment variable
intact) changed the code to require passing an empty list of inputs
rather than None if TEXINPUTS should not be set to a special value. This
broke becaues the compile routine wrapper was not updated to also take
set the default value to be an empty list.

Thanks to David Mehren for reporting this issue.