utils: Improve DPI detection for PDFs that are not a single image
In the case of PDFs that are not a single image the DPI was supposed to
be 300. However, due to rounding issues, the DPI was actually much lower
which was causing detection issues.

Rather than just fixing the rounding issues, try to detect the DPI of
the scan (by looking at all images) and use this DPI. Any unreasonable
DPI will cause a fallback to 300.