model: Improve set_answer handling for Option questions
set_answer just passes in a value. This value needs to be mapped to the
checkboxes for proper storage (as the question itself has no storage
currently, and that would create issues if we e.g. run recognise again).

Change the strategy to set the first box with the correct value. Also
handle the none/invalid special cases by setting no box and all boxes.