model: Number ID and value independently and ensure ID is unique
We had the issue that we bound the ID of boxes to their value. If we do
this and the user overrides a value, then the IDs could get identical,
messing with the internal data store.

For this reason, properly separate the two and redefine the value to be
1 higher than the maximum we have seen so far (starting with 0).

Fixes #142