gui: Mark image as verified on enter and sheet valid if all verified
When pressing Enter, the sheet was being marked as verified. Change this
in the same way as everything else to mark the current image as
verified.

Change the behaviour to mark the sheet as "valid" if the sheet itself is
"verified" (i.e. only after all images have been verified).