tex: Update class for RTL support
Introduces a slight layout change in the choicequestion environment.

With this update, RTL languages should work out of the box. This has not
been extensively tested, and it is a good idea to carefully check the
output of the "annotate" command to verify all checkbox locations are
correct. Errors can be subtle, for example by range/mark questions
having the checkboxes in the wrong order.