Rework command line argument handling
The main difference is that now the project needs to be passed in after
the command that is being executed. The reason to do this is that this
is nicer when adding commands that do not require a project. It also has
the advantage of making the command specific help more easily
accessible.

Note that there are two more minor changes:
 * sdaps_tex as an alias for sdaps has been dropped
 * setting an info key has been changed slightly