image: Fix transformation matrix calculation and move to python
The transformation matrix calculation as originally implemented in C has
been wrong all along. This calculation can be more useful and there is
no need for it to live in C. Also, since it was originally written we
have helper functions to find a specific corner in the C API.

So, add two new functions in python. The first calculate_matrix just
implements the exact same API as the C function does (and overrides it).

The second matrix_from_corners_2d splits out the calculation of the
correction matrix as it can be useful for other purposes.

Fixes: #195