recognize: Add support for non-grouped sheets
In that case, each sheet only contains a single piece of paper (i.e.
two pages of which one is a dummy page in simplex mode).

If that is the case, then inhibit marking the sheet as invalid if a
an image appears to be missing. It is expected to be incomplete after
all.