model: Relax requirement for IDs to always grow
It is OK if they move backwards sometimes in principle. If that happens,
accept the given ID. However, ensure there is no collision and keep
auto-counting at the last increasing value.

Note that the current LaTeX class never uses the auto-counting feature,
so even an undefined behaviour would be acceptable.