recognize: Relax incorrect scan warning rules
If there is a (double) page that does not have any ID information, then
we would emit a warning about an incorrect scan if information is
available for the other ones. This is pretty unlikely to happen, but if
it happens, it does not proof that the scan is incorrect.

The warning is designed to look dangerous (for good reasons), so try
harder to only show it if something is actually wrong, rather than
showing it for a simple bad scan.