paths: Fix build path discovery
We are guessing the build path to run locally. Fix the guessing to work
with newer python versions, falling back to the old behaviour.

Really, this is a very bad hack and needs to be redone/removed. But, not
sure how to best handle it and the whole python build stuff is just not
clear to me.