image: Fix end bitmask for masked pixel counting
Unfortunately, commit git show 2c32e99755f1 ("image: Fix pixel counting
masks for big endian") introduced a regression as it copied the end mask
calculation from the unmasked function, incorrectly adding in the offset
of the area that is to be counted. Remove the incorrect x offset from
the calculation.

Thanks to the Sikshana Foundation for reporting this issue.